+++ src/realm/column_table.cpp
@@ -329,8 +329,7 @@ void SubtableColumn::set(size_t row_ndx, const Table* subtable)
     if (subtable && !subtable->is_empty())
         columns_ref = clone_table_columns(subtable); // Throws
 
-    int_fast64_t value = int_fast64_t(columns_ref);
-    IntegerColumn::set(row_ndx, value); // Throws
+    set_as_ref(row_ndx, columns_ref); // Throws
 
     // Refresh the accessors, if present
     if (Table* table = m_subtable_map.find(row_ndx)) {
@@ -356,6 +355,14 @@ void SubtableColumn::erase_rows(size_t row_ndx, size_t num_rows_to_erase, size_t
         destroy_subtable(row_ndx + i);
 
     SubtableColumnBase::erase_rows(row_ndx, num_rows_to_erase, prior_num_rows, broken_reciprocal_backlinks); // Throws
+}
+
+void SubtableColumn::set_null(size_t row_ndx)
+{
+    REALM_ASSERT_DEBUG(row_ndx < size());
+    destroy_subtable(row_ndx);
+    set_as_ref(row_ndx, 0); // Throws
+    m_subtable_map.adj_set_null(row_ndx);
 }