+++ CHANGELOG.md
@@ -17,6 +17,10 @@
   Query q = table->column<SubTable>(0).list<Int>() == 5;
   Query q = table->column<SubTable>(0).list<Int>().min() >= 2;
   Query q = table->column<SubTable>(1).list<String>().begins_with("Bar");
+  PR [#2532](https://github.com/realm/realm-core/pull/2532)
+* Subtable column can now be nullable. You can use `is_null()` and `set_null()`
+  on a subtable element.
+  PR [#2560](https://github.com/realm/realm-core/pull/2560)
 * Support for search index on subtable columns. Only one level of subtables
   are currently supported, that is, you cannot create a search index in a
   subtable of a subtable (will throw exception). NOTE: Core versions prior to
@@ -24,6 +28,7 @@
   this Core version has added such indexes. Adding or removing an index will
   take place for *all* subtables in a subtable column. There is no way to add
   or remove it form single individual subtables.
+  PR [#2561](https://github.com/realm/realm-core/pull/2561)
 
 -----------